set up app structure
Set up a loose structure for the app including a myutils package, a pages folder, a static folder and a data folder so that styling, functions and data don't have to be written into every page. Set up the calibration page so that it works as the previous version did, but need to add TWIMExtract option and go through AI-written calibrate_page() more carefully.